requirements.txt
tensorflow>=2.13.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
pyyaml>=6.0
google-cloud-bigquery>=3.11.0
google-cloud-storage>=2.10.0
pyarrow>=12.0.0
